Fix issue where property values are cloned
The existing implementation of `ValRefWithSource` forced the
implementation of property access (whether via indexing or direct
property access) cloned the value being accessed, instead of returning a
reference to it. Among other issues, this meant that assigning to values
in nested objects didn't work as expected, because the value would be
assigned to a cloned object. As such, the changes wouldn't persist
relative to the original parent object.